pyproject.toml
[tool.poetry]
name = "phoenix_trader"
version = "0.1.0"
description = ""
authors = ["patelp320 <patelp320@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
yfinance = "^0.2.61"
duckdb = "^1.3.0"
pandas = "^2.2.3"
numpy = "^2.2.6"
scikit-learn = "^1.6.1"
apscheduler = {extras = ["asyncio"], version = "^3.11.0"}
yagmail = "^0.15.293"
ib-insync = "^0.9.86"
aiohttp = "^3.12.6"
feedparser = "^6.0.11"
gitpython = "^3.1.44"
pytest = "^8.3.5"
flake8 = "^7.2.0"
bandit = "^1.8.3"
ollama = "^0.5.1"
requests = "^2.32.3"
joblib = "^1.5.1"
uvloop = "^0.21.0"
python-dotenv = "^1.1.0"
pydantic = "^2.11.5"
typer = "^0.16.0"
litellm = "^1.71.3"
tenacity = "^9.1.2"
backoff = "^2.2.1"
toml = "^0.10.2"
tomli = "^2.2.1"
termcolor = "^3.1.0"
playwright = "^1.52.0"
json-repair = "^0.46.0"
chromadb = "^1.0.12"
llama-index = "^0.12.39"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
